test(alert): migrate isOpen
